+++ apps/swirl-docs/src/scripts/prebuild.ts
@@ -227,7 +227,6 @@ function deleteGlobalSpecs() {
   deleteSpec("version-info.yml");
   deleteSpec("merged.yml");
   deleteSpec("organisations.yml");
-  deleteSpec("posts.yml");
   deleteSpec("sharepoint-pages.yml");
 }