+++ packages/swirl-tokens/dart/lib/styles.dark.dart
@@ -4,7 +4,11 @@
 //
 
 // Do not edit directly
+<<<<<<< HEAD
 // Generated on Tue, 29 Aug 2023 10:01:25 GMT
+=======
+// Generated on Mon, 09 Oct 2023 10:13:22 GMT
+>>>>>>> main